Split some code up into multiple files
