+++ tests/test_speedrun_demorecord.py
@@ -15,245 +15,21 @@ Future Tests:
 * Unicode paths & handling
 """
 
-import pytest
-import os
-import subprocess
-import shutil
-import re
 import glob
+import os
+import re
+import shutil
+import subprocess
+from typing import Dict, Iterable, List, Optional, cast
 
-from enum import IntEnum, auto
-from typing import Optional, List, Dict, cast, Iterable
+import pytest
 
-# ----- TODO REMOVE THIS WHEN DONE DEBUGGING ---------
-os.environ[
-    "GAME_2013_HL2_DIR"] = "C:\\Program Files (x86)\\Steam\\steamapps\\common\\Half-Life 2"
-os.environ["GAME_2007_HL2_DIR"] = "C:\\SourceUnpack\\hl2_vanilla"
-# ----- TODO REMOVE THIS WHEN DONE DEBUGGING ---------
-
-
-class SourceEngineGame(object):
-    def __init__(self, game_tag: str, game_dir_env_var: str, exe_filename: str,
-                 game_short_code: str, test_data_dir: str,
-                 expected_demo_names: List[str]):
-        resolved_game_dir: str = os.path.abspath(
-            os.getenv(game_dir_env_var, ""))
-        self.__game_tag: str = game_tag
-        self.__game_dir: str = os.path.join(resolved_game_dir, game_short_code)
-        self.__exe_path: str = os.path.join(resolved_game_dir, exe_filename)
-        self.__game_short_code: str = game_short_code
-        self.__test_data_dir: str = test_data_dir
-        self.__expected_demo_names: List[str] = expected_demo_names
-
-        # Parse layout config
-        self.__test_data_file_list: List[str] = []
-        self.__test_data_playback_cfg: str = ""
-        self.__parse_layout_config()
-
-    @property
-    def game_tag(self) -> str:
-        return self.__game_tag
-
-    @property
-    def exe_path(self) -> str:
-        return self.__exe_path
-
-    @property
-    def game_dir(self) -> str:
-        return self.__game_dir
-
-    @property
-    def launch_args(self) -> List[str]:
-        return [
-            self.__exe_path, '-game', self.__game_short_code, '-novid',
-            '-allowdebug', '-sw', '-w', '1280', '-h', '720'
-        ]
-
-    @property
-    def expected_demo_names(self) -> List[str]:
-        return self.__expected_demo_names
-
-    @property
-    def test_data_dir(self) -> str:
-        return self.__test_data_dir
-
-    @property
-    def test_data_file_list(self) -> List[str]:
-        return self.__test_data_file_list
-
-    @property
-    def test_data_playback_cfg_path(self) -> str:
-        return os.path.relpath(self.__test_data_playback_cfg, 'cfg/')
-
-    def __parse_layout_config(self) -> None:
-
-        layout_path: str = os.path.join(self.__test_data_dir, 'layout.ini')
-        if not os.path.isfile(layout_path):
-            return
-
-        file_paths: List[str] = []
-        with open(layout_path, 'r', encoding='utf-8') as fd:
-            for line in fd:
-
-                # Ignore comments
-                normalized_line: str = line.strip()
-                if not normalized_line.startswith('#'):
-                    k_v_pair: List[str] = normalized_line.split('=')
-                    assert len(k_v_pair) == 2
-
-                    # Ensure file actually exists if it isn't a glob path
-                    rel_file_path: str = k_v_pair[1]
-
-                    abs_file_path: str = os.path.join(self.__test_data_dir,
-                                                      rel_file_path)
-                    if not glob.has_magic(
-                            abs_file_path) and not os.path.isfile(
-                                abs_file_path):
-                        raise Exception(
-                            f"the file \"{abs_file_path}\" does not exist as defined in \"{layout_path}\"."
-                        )
-
-                    # Playback config
-                    if k_v_pair[0] == 'playback':
-                        assert not self.__test_data_playback_cfg
-                        self.__test_data_playback_cfg = rel_file_path
-
-                    file_paths.append(rel_file_path)
-        self.__test_data_file_list = file_paths
-
-
-SUPPORTED_GAMES: List[SourceEngineGame] = [
-    SourceEngineGame(game_tag='hl2_2013',
-                     game_dir_env_var='GAME_2013_HL2_DIR',
-                     exe_filename='hl2.exe',
-                     game_short_code='hl2',
-                     test_data_dir=os.path.join(os.path.dirname(__file__),
-                                                'reproduction', 'hl2_2013'),
-                     expected_demo_names=[
-                         "d1_canals_06.dem", "d1_canals_06_1.dem",
-                         "d1_canals_06_2.dem", "d1_canals_07.dem",
-                         "d1_canals_06_3.dem", "d1_canals_06_4.dem",
-                         "d1_canals_08.dem", "d2_coast_01.dem"
-                     ]),
-    SourceEngineGame(game_tag='hl2_2007',
-                     game_dir_env_var='GAME_2007_HL2_DIR',
-                     exe_filename='hl2.exe',
-                     game_short_code='hl2',
-                     test_data_dir=os.path.join(os.path.dirname(__file__),
-                                                'reproduction', 'hl2_2007'),
-                     expected_demo_names=[
-                         "d1_canals_06.dem", "d1_canals_06_1.dem",
-                         "d1_canals_06_2.dem", "d1_canals_07.dem",
-                         "d1_canals_06_3.dem", "d1_canals_06_4.dem",
-                         "d1_canals_08.dem", "d2_coast_01.dem"
-                     ])
-]
+from demo_utils import construct_vdm, get_demo_tick_count
+from source_engine_games import SUPPORTED_GAMES, SourceEngineGame
 
 RE_EXPECTED_DATETIME_DIR = re.compile(
     r"[0-9]{4}\.[0-9]{2}\.[0-9]{2}-[0-9]{2}\.[0-9]{2}\.[0-9]{2}")
 SPEEDRUN_DEMORECORD_DEMO_FOLDER: str = "./integration_tests/"
-
-
-class DemMsgType(IntEnum):
-    Nop = 0
-    Signon = auto()
-    Packet = auto()
-    SyncTick = auto()
-    ConsoleCmd = auto()
-    UserCmd = auto()
-    DataTables = auto()
-    Stop = auto()
-    StringTables = auto()
-
-
-DEM_MSG_DATA: List[DemMsgType] = [
-    DemMsgType.Signon, DemMsgType.Packet, DemMsgType.ConsoleCmd,
-    DemMsgType.UserCmd, DemMsgType.DataTables, DemMsgType.StringTables
-]
-DEM_MSG_OTHER: List[DemMsgType] = [DemMsgType.SyncTick, DemMsgType.Nop]
-DEM_HEADER_SIZE = 0x430
-
-
-def get_demo_tick_count(demo_path: str) -> Optional[int]:
-    with open(demo_path, 'rb') as fd:
-        # Grab firt eight bytes for header start confirmation
-        demo_header_start_bytes: bytes = fd.read(8)
-
-        # Assert header correct
-        assert demo_header_start_bytes == b"HL2DEMO\x00"
-
-        # Move past header
-        fd.seek(DEM_HEADER_SIZE)
-
-        tick: Optional[int] = None
-
-        while True:
-            dem_msg: int = int.from_bytes(fd.read(1),
-                                          byteorder='little',
-                                          signed=True)
-
-            if dem_msg == DemMsgType.Stop:
-                break
-
-            tmp_tick = int.from_bytes(fd.read(4),
-                                      byteorder='little',
-                                      signed=True)
-            if tmp_tick >= 0:
-                # Ignore negative tick values
-                tick = tmp_tick
-
-            if dem_msg in DEM_MSG_DATA:
-                if dem_msg == DemMsgType.Packet or dem_msg == DemMsgType.Signon:
-                    fd.read(0x54)
-                elif dem_msg == DemMsgType.UserCmd:
-                    fd.read(0x4)
-
-                data_len: int = int.from_bytes(fd.read(4), byteorder='little')
-                fd.read(data_len)
-            elif dem_msg in DEM_MSG_OTHER:
-                continue
-            else:
-                raise Exception(f"Unknown message {dem_msg}!")
-
-        # Return actual tick count
-        return tick
-
-
-def construct_vdm(demo_path: str, next_demo: Optional[str]):
-    # next_demo should be a path relative to the game folder (?)
-    demo_endtick: Optional[int] = get_demo_tick_count(demo_path)
-    if not demo_endtick:
-        raise Exception("Failed to parse end tick from demo \"{demo_path}\"")
-
-    # Remove two ticks just to make sure the PlayCommand action works
-    demo_endtick -= 2
-    demo_file_path: str = os.path.splitext(demo_path)[0]
-    demo_vdm_path: str = f"{demo_file_path}.vdm"
-
-    cmd_at_end: str = "wait 200; "
-    if next_demo is None:
-        cmd_at_end += "exit"
-    else:
-        cmd_at_end += f"playdemo {next_demo}"
-
-    with open(demo_vdm_path, 'w', encoding='utf-8') as fd:
-        fd.write(f"""demoactions
-{{
-    "1"
-    {{
-        factory "PlayCommands"
-        name "Execute command"
-        starttick "{demo_endtick}"
-        commands "{cmd_at_end}"
-    }}
-    "2"
-    {{
-        factory "ChangePlaybackRate"
-        name "Play demo fast"
-        starttick "0"
-        playbackrate "10.000000"
-    }}
-}}""")
 
 
 @pytest.fixture
